create `selectSvgElement`
